Add more exclusions to the backup

+++ bin/backup.sh
@@ -6,7 +6,7 @@ declare -r WORKSPACE="${HOME}/workspace"
 declare -r DESTINATION_DIRECTORY="${HOME}/Dropbox/Backups/IBM"
 
 DATE=$(date +${DATE_FORMAT})
-DIRECTORIES=(go src linuxacademy ibm/src ibm/infrastructure clarity/documentation)
+DIRECTORIES=(acloudguru codelytv go ibm/infrastructure ibm/src kike linuxacademy micifuz mundokids src)
 
 
 command::do_backup( ) {
@@ -14,10 +14,23 @@ command::do_backup( ) {
   local complete_directory="${WORKSPACE}/${directory}"
   local tar_file_name="/tmp/${directory//\//-}-${DATE}.tar.gz"
 
-  [[ -d "${complete_directory}" ]] && \
-    ${TAR} -czvf ${tar_file_name} --exclude='.git' --exclude='.DS_Store' --exclude='.terraform' --exclude='.terragrunt-cache' --exclude='vendor' --exclude='pkg/mod' --exclude='node_modules' --exclude='.gradle' --exclude='*.class' --exclude='*.jar' --exclude='*.csv' --exclude='.code' --exclude='.idea' --exclude=.po --exclude=__pycache__ -C "${complete_directory}" . && \
-    cp "${tar_file_name}" "${DESTINATION_DIRECTORY}/" && \
+  if [ -d "${complete_directory}" ]; then
+    ${TAR} -czf ${tar_file_name} \
+      --exclude='.git'\
+      --exclude='.DS_Store' \
+      --exclude='.terraform' \
+      --exclude='.terragrunt-cache' \
+      --exclude='vendor' \
+      --exclude='pkg/mod' \
+      --exclude='node_modules' \
+      --exclude='.gradle' --exclude='*.class' --exclude='*.jar' \
+      --exclude='*.csv' \
+      --exclude='.code' --exclude='.idea' \
+      --exclude='.po' --exclude='.pc' --exclude='__pycache__' \
+      -C "${complete_directory}" . && \
+    cp "${tar_file_name}" "${DESTINATION_DIRECTORY}/" &&
     rm -f "${tar_file_name}"
+  fi
 }
